Disable persistence of ReportDraft in localStorage
There are a bunch of Sentry bugs where some fields of the ReportDraft are missing. I'm thinking it's consistency problem between all the load()/save() performed on the localStorage, at some point some stale data must be loaded and we lose some fields.
Let's try to disable entirely the save in localstorage for the report draft, to see if the Sentry bugs get better.